Input: aw8624_haptic - fix printk formatting

+++ drivers/input/misc/aw8624_haptic/aw8624.c
@@ -2794,8 +2794,7 @@ static int aw8624_parse_dt(struct device *dev, struct aw8624 *aw8624,
 		}
 		printk
 		    ("%s: %d  effect->pattern_length=%d  effect->pattern=%d \n",
-		     __func__, __LINE__, effect->pattern_length,
-		     (int)effect->pattern);
+		     __func__, __LINE__, effect->pattern_length, effect->pattern);
 
 		effect->play_rate_us = config->play_rate_us;
 		rc = of_property_read_u32(child_node, "qcom,wf-play-rate-us",
